debug

+++ sql/chat.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS `chat`;
 CREATE TABLE `chat`
 (
-    `id`           int          NOT NULL UNIQUE AUTO_INCREMENT COMMENT 'id',
-    `from_user_id` int          NOT NULL COMMENT '发送用户id',
-    `to_user_id`   int          NOT NULL COMMENT '接收消息用户id',
+    `id`           bigint          NOT NULL UNIQUE AUTO_INCREMENT COMMENT 'id',
+    `from_user_id` bigint          NOT NULL COMMENT '发送用户id',
+    `to_user_id`   bigint          NOT NULL COMMENT '接收消息用户id',
     `content`      varchar(300) NOT NULL,
     `create_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP ,
     `update_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP ,
