+++ sql/like.sql
@@ -1,9 +1,9 @@
 DROP TABLE IF EXISTS `like`;
 CREATE TABLE `like`
 (
-    `id`          int        NOT NULL UNIQUE AUTO_INCREMENT COMMENT 'id',
-    `user_id`     int        NOT NULL COMMENT '用户id',
-    `video_id`  int        NOT NULL COMMENT '点赞视频id',
+    `id`          bigint        NOT NULL UNIQUE AUTO_INCREMENT COMMENT 'id',
+    `user_id`     bigint        NOT NULL COMMENT '用户id',
+    `video_id`  bigint        NOT NULL COMMENT '点赞视频id',
     `create_time` timestamp  NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT '创建时间',
     `update_time` timestamp  NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT '更新时间',
     PRIMARY KEY (`id`),
